mcp/mcp-cyber/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-cyber"
version = "0.1.0"
description = "Enterprise defensive security validation platform (allowlist-gated scans, MCP + API)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.6",
    "mcp>=1.0.0",
    "pyyaml>=6.0.1",
    "psycopg[binary]>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cyber_*"]

[project.scripts]
mcp-cyber-mcp = "cyber_mcp.server:main"
mcp-cyber-api = "cyber_api.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
